backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<26.0
psycopg[binary]>=3.2,<3.3
django-cors-headers>=4.6,<5.0
python-dotenv>=1.0,<2.0
Pillow>=11.0,<13.0
celery>=5.4,<6.0
redis>=5.2,<7.0
